mindbridge/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==23.2.1
pydantic==2.7.1
